Major Refactoring ( Object Pooling + Health)

+++ core_mechanics_challenge_winter-2025/Assets/Scripts/Player/PlayerStats.cs
@@ -4,7 +4,7 @@ public class PlayerStats : MonoBehaviour
 {
     public float MoveSpeed = 2f;
     public float FireRate = 0.5f;
-    public float Damage = 1f;
+    public int Damage = 1;
 
     public float BulletSize = .2f;
     public int BulletPierce = 1;
@@ -15,7 +15,7 @@ public class PlayerStats : MonoBehaviour
 
     public const float MAX_MOVE_SPEED = 5f;
     public const float MIN_FIRE_RATE = 0.1f;
-    public const float MAX_DAMAGE = 10f;
+    public const int MAX_DAMAGE = 10;
     public const float MAX_BULLET_SIZE = .6f;
     public const int   MAX_PIERCE = 5;
 }
